# Incorporate plugin health score in JSON metadata