--- a/src/main/java/io/jenkins/update_center/GitHubSource.java
+++ b/src/main/java/io/jenkins/update_center/GitHubSource.java
@@ -108,8 +108,7 @@ public class GitHubSource {
                 );
             }
             HttpRequest request = builder.build();
-            try {
-                final HttpClient client = HttpClient.newHttpClient();
+            try (final HttpClient client = HttpClient.newHttpClient()) {
                 final HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
                 final JSONObject jsonResponse = JSONObject.fromObject(response.body());
                 if (jsonResponse.has("errors")) {