--- a/src/main/java/io/jenkins/update_center/HealthScores.java
+++ b/src/main/java/io/jenkins/update_center/HealthScores.java
@@ -3,6 +3,8 @@ package io.jenkins.update_center;
 import java.io.IOException;
 import java.util.Map;
 import java.util.function.Function;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 import java.util.stream.Collectors;
 
 import com.alibaba.fastjson.JSON;
@@ -11,6 +13,7 @@ import okhttp3.OkHttpClient;
 import okhttp3.Request;
 
 public class HealthScores {
+    private static final Logger LOGGER = Logger.getLogger(HealthScores.class.getName());
     private static final String HEALTH_SCORES_URL = "https://reports.jenkins.io/plugin-health-scoring/scores.json";
 
     private static HealthScores instance;
@@ -20,24 +23,29 @@ public class HealthScores {
         this.healthScores = healthScores;
     }
 
-    public static synchronized HealthScores getInstance() throws IOException {
+    public static synchronized HealthScores getInstance() {
         if (instance == null) {
             initialize();
         }
         return instance;
     }
 
-    private static void initialize() throws IOException {
+    private static void initialize() {
         final Request request = new Request.Builder().url(HEALTH_SCORES_URL).get().build();
-        final String bodyString = HttpHelper.getResponseBody(new OkHttpClient(), request);
 
-        final JsonResponse response = JSON.parseObject(bodyString, JsonResponse.class);
-        if (response.plugins == null) {
-            throw new IllegalArgumentException("Specified popularity URL '" + HEALTH_SCORES_URL + "' does not contain a JSON object 'plugins'");
+        try {
+            final String bodyString = HttpHelper.getResponseBody(new OkHttpClient(), request);
+            final JsonResponse response = JSON.parseObject(bodyString, JsonResponse.class);
+            if (response.plugins == null) {
+                throw new IllegalArgumentException("Specified popularity URL '" + HEALTH_SCORES_URL + "' does not contain a JSON object 'plugins'");
+            }
+
+            final Map<String, Integer> healthScores = response.plugins.keySet().stream().collect(Collectors.toMap(Function.identity(), pluginId -> response.plugins.get(pluginId).value));
+            instance = new HealthScores(healthScores);
+        } catch (IOException e) {
+            LOGGER.log(Level.WARNING, e.getMessage());
+            instance = new HealthScores(Map.of());
         }
-
-        final Map<String, Integer> healthScores = response.plugins.keySet().stream().collect(Collectors.toMap(Function.identity(), pluginId -> response.plugins.get(pluginId).value));
-        instance = new HealthScores(healthScores);
     }
 
     private static class JsonResponse {