… pom.xml
… src/main/java/io/jenkins/update_center/Popularities.java
… src/main/java/io/jenkins/update_center/util/HttpHelper.java
… src/test/java/io/jenkins/update_center/WarningsTest.java