This adds the plugin health, from the https://github.com/jenkins-infra/plugin-health-scoring project, into the update-center output.

This is the first step to be able to display the scores in the Plugin Manager.

Corresponding core PR: https://github.com/jenkinsci/jenkins/pull/10351

# Testing

I tested on a limited number of plugins and it worked. I'm testing now with all plugins.