@alecharp 2025-02-03
> * What is the impact on the size of the JSON metadatas? I have vague recollections of impacts due to adding even an integer field causing challenges. Of course it might be hard to evaluate: if it is the case, I would propose to start with generating both JSONs files (with and without the PH score), we keep deploying the "without PH score" and measure the impact in size and build/deploy time before going forward.

I'll generate both and let you down the time required for the generation and the size the the files in each case.

> * This PR adds a dependency on the service reports.jenkins.io. If the service is down or not reachable, what would be the expected fall backs? This webservice is "HA" (e.g. it has a QoS bound to the underlying Azure Kubernetes cluster today) but could be down.

Technically, we already have this dependency: 
https://github.com/jenkins-infra/update-center2/blob/master/src/main/java/io/jenkins/update_center/MaintainersSource.java#L47-L48. However, I'll try to see what happens and improve the code if needed for this situation.