@dduportal 2025-02-03
> I'll generate both and let you down the time required for the generation and the size the the files in each case.

I wasn't sure it was possible or easy for you (as the production dataset could be heavy, not even speaking about the cache...). Thanks!

> Technically, we already have this dependency: https://github.com/jenkins-infra/update-center2/blob/master/src/main/java/io/jenkins/update_center/MaintainersSource.java#L47-L48. However, I'll try to see what happens and improve the code if needed for this situation.


OH: I was not aware. Then your change here does NOT add any dependency but reuse existing. It makes sense then (as per my infra hat: I can answer for the scope of the code itself).