@alecharp 2025-02-03
So for the size:

```bash
❯ ls -al with*
with-health-score:
total 18528
drwxr-xr-x   5 alecharpe  *      160  3 fév 15:25 .
drwxr-xr-x  20 alecharpe  *      640  3 fév 15:28 ..
-rw-r--r--   1 alecharpe  *  3158562  3 fév 15:25 update-center.actual.json
-rw-r--r--   1 alecharpe  *  3158584  3 fév 15:25 update-center.json
-rw-r--r--   1 alecharpe  *  3158787  3 fév 15:25 update-center.json.html

without-health-score:
total 18312
drwxr-xr-x   5 alecharpe  *      160  3 fév 15:28 .
drwxr-xr-x  20 alecharpe  *      640  3 fév 15:28 ..
-rw-r--r--   1 alecharpe  *  3124324  3 fév 15:28 update-center.actual.json
-rw-r--r--   1 alecharpe  *  3124346  3 fév 15:28 update-center.json
-rw-r--r--   1 alecharpe  *  3124549  3 fév 15:28 update-center.json.html
```

For the timing, I ran the commands with an existing cache as downloading the files should not be part of the problem. As the health report is accessed / loaded once, this changeset should not be the biggest source of issues. Compared to accessing all the plugin GitHub repositories for example.

* without the health score: `java -Dfile.encoding=UTF-8 -jar  --id default --www-dir without-health-score  14,61s user 1,77s system 14% cpu 1:55,94 total`
* with the health score: `java -Dfile.encoding=UTF-8 -jar  --id default --www-dir with-health-score  13,82s user 1,71s system 12% cpu 2:02,35 total`

I'm still looking at what happens with no access to reports service.