@alecharp 2025-02-03
So apparently, with reports.jenkins.io is down (blocked it locally), the `master` branch is not handling that well: it breaks the generation entirely. 
I don't think there is a point of testing this scenario for this changeset. But that's something we can track to improve in the future.