@alecharp 2025-02-03
Thanks @daniel-beck. Just improved the code to not fail if the report is not available.
I would like to take some time to see if we could remove the health score integer completely from the generated file if we didn't get a score. Cause, if we provide a 0 because reports service is down, that seems bad.