@daniel-beck 2025-02-04
On-holding this, pending the corresponding core change.