@alecharp 2025-02-25
It now also remove OkHTTP dependency requirement.