@daniel-beck 2025-04-11
I switched generation of update sites to a snapshot uploaded from this branch, so currently, real update sites have `healthScore`.